pyproject.toml
[project]
name = "kindred-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "sqlacodegen>=3.0.0",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.2",
]
